# improve `heatmap_edges` performance