Partial improvement for https://github.com/JuliaPlots/Plots.jl/issues/4520.

```julia
using BenchmarkTools, Plots

const mat = randn(1024, 1024);

@btime show(devnull, heatmap($mat))
  35.233 ms (632 allocations: 8.10 MiB)  # PR
  35.512 ms (4670 allocations: 8.25 MiB)  # master

f(x, n = 100) = begin
    plot()
    for _ in 1:n
        heatmap!(x)
    end
end

@btime f($mat)
  2.320 s (42563 allocations: 807.30 MiB)  # PR
  2.348 s (446363 allocations: 822.83 MiB)  # master
```